NLP_Projects/image_captioning/requirements.txt.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
spacy>=3.1.0
tqdm>=4.62.0
matplotlib>=3.4.0
nltk>=3.6.0
